refactor: - Se incrementa la versión del proyecto para reflejar ajustes en la estructura de responsabilidades. - Cambios incluyen: mover NavigationContext a application/context, mantener LoadingContext en shared/context y mejorar modularidad, mover hooks de podcast a casos
