+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "podcaster",
-  "version": "7.0.4",
+  "version": "7.1.0",
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^6.6.3",
